- DOC: config.php - BUG: testmode did not check if sitemap.xml existed - BUG: DB test on db errors with new _set_key_to()
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@3298 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

Former-commit-id: c0d26191eae8ef4c06f453f9c6031e582a6efb25
Former-commit-id: 42fd4cbfe97d3288df72ed769741683a4748ddee
Former-commit-id: 0bd56bf1c1a9b650f059ab9baf67a1475d1384b3 [formerly 753a0821304ab010290502b46fb1369f479d6f7b]
Former-commit-id: 0b9ca695255c855244cbb4c6fa379acda4d461d4

+++ sys/flexyadmin/core/FrontendController.php
@@ -51,7 +51,7 @@ class FrontEndController extends MY_Controller {
       if (!$this->user->is_super_admin()) {
         $temp=file_get_contents('index_temp.html');
         echo $temp;
-        unlink('sitemap.xml');
+        if (file_exists('sitemap.xml')) unlink('sitemap.xml');
         die();
       }
     }
